requirements.txt
pywin32
pystray
Pillow
requests
schedule
customtkinter
